Fix current_time and put all the time getting code in one place (#200)
* put all the time getting code in one place

* comment

* add stuff

* remove platformGetTime()

* nit

* remove getTime

* less double math

* put it in the runner struct

* set thingy

* set

* fix

---------

+++ src/desktop/platformdefs.h
@@ -17,7 +17,7 @@ bool platformGetWindowSize(int32_t* outW, int32_t* outH);
 bool platformGetScaledWindowSize(int32_t* outW, int32_t* outH);
 void platformSetWindowSize(int32_t width, int32_t height);
 void platformSetWindowTitle(const char* title);
-void platformSleepUntil(double time);
+void platformSleepUntil(uint64_t time);
 
 enum GraphicsAPI {
     SOFTWARE,
